+++ src/docs/ChangeLog.md
@@ -11,6 +11,7 @@
 * All events are namespaced properly `{event}.datetimepicker`
 * Added a jquery no conflict option
 * Removed LESS files. Bootstrap 4 doesn't use LESS any more. Now using SASS to build CSS
+* Picker will also look for window.debug and will keep the picker from closing
 
 ### Other changes